+++ ISO4217Test.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * (c) Rob Bast <rob.bast@gmail.com>
  *
@@ -16,13 +18,11 @@ class ISO4217Test extends TestCase
     /**
      * @testdox Calling getByAlpha3 with an invalid alpha3 throws a DomainException.
      *
-     * @param string|int $alpha3
      * @dataProvider invalidAlpha3Provider
      */
-    public function testGetByAlpha3Invalid($alpha3): void
+    public function testGetByAlpha3Invalid($alpha3, $expectException): void
     {
-        $this->expectException(\DomainException::class);
-        $this->expectExceptionMessageMatches('/^Not a valid alpha3: .*$/');
+        $this->expectException($expectException);
 
         $iso4217 = new ISO4217();
         $iso4217->getByAlpha3($alpha3);
@@ -52,13 +52,11 @@ class ISO4217Test extends TestCase
 
     /**
      * @testdox Calling getByNumeric with an invalid numeric throws a DomainException.
-     *
      * @dataProvider invalidNumericProvider
      */
-    public function testGetByNumericInvalid(string $numeric): void
+    public function testGetByNumericInvalid($numeric, $expectException): void
     {
-        $this->expectException(\DomainException::class);
-        $this->expectExceptionMessageMatches('/^Not a valid numeric: .*$/');
+        $this->expectException($expectException);
 
         $iso4217 = new ISO4217();
         $iso4217->getByNumeric($numeric);
@@ -98,7 +96,14 @@ class ISO4217Test extends TestCase
 
     public function invalidAlpha3Provider(): array
     {
-        return [['ZZ'], ['ZZZZ'], [12], [1234]];
+        return [
+            ['00', \DomainException::class],
+            ['0000', \DomainException::class],
+            ['ZZ', \DomainException::class],
+            ['ZZZZ', \DomainException::class],
+            [12, \TypeError::class],
+            [1234, \TypeError::class],
+        ];
     }
 
     public function alpha3Provider(): array
@@ -108,7 +113,14 @@ class ISO4217Test extends TestCase
 
     public function invalidNumericProvider(): array
     {
-        return [['00'], ['0000'], ['ZZ'], ['ZZZZ']];
+        return [
+            ['00', \DomainException::class],
+            ['0000', \DomainException::class],
+            ['ZZ', \DomainException::class],
+            ['ZZZZ', \DomainException::class],
+            [12, \TypeError::class],
+            [1234, \TypeError::class],
+        ];
     }
 
     public function numericProvider(): array